feat: store remote-envs persistently and centrally (#901)
* feat: store remote-envs persistently and centrally

store a single copy for each remote env in
`$FLOX_CACHE_DIR/remote/<owner>/<name>`

* fix: reduce unnecessary rebuilds of remote and managed environments

Pulling a managed environment _always_ causes a rebuild even if no
changes were pulled.
That is because each pull would (re) write the generation lockfile at
`.flox/env.lock`.
Since rebuilding depends on the mtime of that file, we rebuilt
unnecessarily often.

The solution for now is to skip writing the file iff
(1) the file exists
(2) the contents are unchanged

* feat: protect active environment state until push succeded 

The inner managed environment of a remote environment is first updated,
_and then_ pushed.
If the push fails $FLOX_ENV will point to the updated but unpushed
out_link of the inner environment.

RemoteEnvironment now refers to a separate `out_link`,
which is a symlink to the store path of the inner managed environment.
When the push succeeds, that link is updated to point to the new store
path.

---------

Co-authored-by: Matthew Kenigsberg <matthew@floxdev.com>