build: make package for flox-manpages (#916)
Manpages were previously built as part of the `flox-cli` package, which
placed them in a `$man` output.
`symlinkJoin` only linked the default `$out` output.
Hence, manpages were lost in the final build.

This PR separates the build of the manpages into it's own package and
`symlinkJoins` both `flox-cli` and `flox-manpages` into the final build.

closes #909