feat: add `Message` abstraction (#947)
> **tl;dr:** 
>
> This PR should make the messaging across `flox` more consistent
> If we provide _messages_ to users (success, error, warning, creation,
update,...)
> We can do so through the new `message` module introduced here.
> Lifting this responsibility from the logging subsystem,
> which can now be used for its intended purpose, logging --
> optional semi user facing (structured) messages.


We are currently using a mix of `println!`, `eprintln!`, `log::info!`,
`log::error!`
to display _messages_ to users on the CLI.

* `println!` => processable output.
* `log::error!` => error messages, prefixed with `ERROR:`
* `eprintln!` and `log::info` => other user facing messages without or
limited processing

Using `log` for user output is suboptimal since `log` will do double
duty
for logging and messaging.
While `log::error` applied styling and was used semantically correct,
where we wanted to express _errors_, `log::info` is overloaded for all
kinds of messages
and thus requires styling to be applied by the caller.

The `message` module now provides explicit functions for several kinds
of user facing messages, making it a nudging towards deliberate use of
message styles in accordance with the style guide.