fix: do not copy user config into test suite
When setting up the test suite, we used to _copy_ the user's nix and flox config
into the otherwise isolated managed test suite.
The commentary on the containing function suggests that _caches_
are copied for efficiency.
In the absence of compelling reason to impurely use user config files in the test suite,
this removes the section that copies flox and nix config dirs and replaces it with
`mkdir -p`.