fix: ld-floxlib compatiblity, re-enable test (#978)
- Revert skipping ld-floxlib test. The test appears to have been failing
because bats was backgrounding a `flox activate -- ...` command. At the
time `flox activate` invoked `bash -i -c`. If a background process tries
to read from stdin it is sent SIGTTIN. We've since merged
https://github.com/flox/flox/pull/889 which no longer invokes `bash`
with `-i`, so it's no longer attempting to read from stdin
- Build ld-floxlib for default make target. Without this change,
ld-floxlib is not built by `just integ-tests`
- Update `ld-floxlib.c` to "pin" GLIBC symbol versions to the minimum
available (varies by architecture type) so as to maximise compatibility
with older versions of GLIBC.
- Update `ld-floxlib.c` to replace the use of `stat()` with `open()`,
which is a) a function symbol in glibc that can be bound to a specific
version (unlike `stat()`) and b) able to assess both the existence and
the "readability" of a given library in a single step, which is
something that `stat()` did not do.
- Continue invoking `patchelf` to remove the runpath from
`ld-floxlib.so`, but do it from within the Makefile target rather than
post-processing the binary in the Nix expression following the build.