chore: remove `flox-types` crate
Remove separate crate in favor of a `data` module in `flox-rust-sdk`.
`flox-types` was meant to collect "plain data" types,
of which only `System` and `Version` remain in use currently.
This change moves the latter to a new `data` module in `flox-rust-sdk`.
Other plain data structs currently defined in `flox-rust-sdk` may be moved in the future.