fix: independent runs of search and show tests (#1103)
Before this change, the search and show tests were depending on state in
the global lockfile, and this would cause
`just integ-file package-show.bats` to fail. Instead, explicitly force
the nixpkgs rev tested.

Also remove project_setup and project_teardown where they are not
necessary and run search and show tests in the global context.