feat: use rules for scraping. (#1095)
## Proposed Changes

Add allow/disallow recurse/no-recurse rules for scraping. This PR should
not change the scope of what is scraped. Those changes (additional
rules) will be added as separate PRs. Notes from the original PR have
been carried over.

- Adds the ability to provide `pkgdb` scrape creation rules.
- Rules can be edited in a hard coded `pkgdb/src/pkgdb/rules.json.hh`.

Notable limitations:
- Cannot nest `allow*` under `disallowRecursive` ( docs in
`rules.json.hh` reflect this limitation ).
+ `applyRules()` - supports nesting `allow*` under `disallowRecursive`.
  + `scrape()` - implementation may prevent this from working.
- Hard coded rules file - previous PR allowed for a command line
argument to define the rules file, however, it's not clear when or if
that will be needed and had several `FIXME`s in the implementation.
Removed on YAGNI principles.
- No database invalidation when rules change.
+ It is strongly recommended that some hash ( similar to fingerprint )
be added to a meta fields of DBs to invalidate them if and when a
command line argument is provided for the rules to use.
  + For now, delete your caches during local dev between rule changes.

## Release Notes

N/A