feat: add nodejs init hooks (#1105)
Implement nodejs init hooks, but only run them if _FLOX_NODE_HOOK is
set, since the npm and yarn hooks are not yet implemented.

- Look for nodejs version in .nvmrc
- Look for nodejs version in the engines.node field of packages.json

Depending on whether a version is found and whether Flox can provide it,
offer to install nodejs.

If offering to install nodejs, look for package-lock.json and yarn.lock,
and depending on which are present, offer to add a hook.