feat: install npm or yarn in node `flox init` hook
Depending on which of package.json, package-lock.json, and yarn.lock are
present, check for compatible versions of npm and yarn.

npm and yarn hardcode nixpkgs#nodejs as the node they use, so make sure
nixpkgs#nodejs is compatible engines.node before offering to install
either npm or yarn.

If constraints are not satisfied, fall back to just installing nodejs
without npm or yarn.

The node init hook is left feature flagged with _FLOX_NODE_HOOK.