feat: install yarn in node `flox init` hook
Depending on which of package.json, package-lock.json, and yarn.lock are
present, check for a compatible version of yarn.

yarn hardcodes nixpkgs#nodejs as the node it uses, so make sure
nixpkgs#nodejs is compatible with engines.node before offering to
install yarn.

If constraints are not satisfied, fall back to just installing nodejs.

At this point, don't check engines.npm, because nixpkgs only contains
one version of npm, but it contains more versions of npm bundled in with
the several different versions of nodejs it contains.

The node init hook is left feature flagged with _FLOX_NODE_HOOK.