fix: set NIX_SSL_CERT_FILE for in pkgdb
pkgdb calls nix functions that require an SSL cert to download flakes

We previously set this env var for the entirety of the flox process, so
all children would have the variable set as well.

Commit 46c49495e0847f49d5955087b2a3557de163c791 removed:
```
// Set the global Nix config via the environment variables in flox.default_args so that
// subprocesses called by `flox` (e.g. `parser-util`) can inherit them.
flox.nix::<NixCommandLine>(vec![]).export_env_vars();
```

So we now need to set this variable explicitly in pkgdb.