fix: implement parent shell detection by process id (#1176)
We currently rely on `$SHELL` for our detection of the executing shell.
However, sub shells do not reliably reset this variable.
When `flox activate --print-script` is `eval`'ed in a sub shell
-- e.g. a direnv issued bash -- an incompatible script may produced.

    SHELL=zsh bash -c 'eval "$(flox activate --print-script)"'
    .../activate/zsh:13: setopt: command not found
    .../activate/zsh:14: setopt: command not found

This PR implements detection of the parent shell by determining
and inspecting the exe path of the parent process.
If determining the parent process fails,o r the process is an
unsupported shell,
we fall back to the existing `$SHELL` detection.

Both methods are overridden by an explicit `FLOX_SHELL` variable.