fix: wrap pkgdb initialization in an EXCLUSIVE transaction
As soon as we have a connection, we need
1) create the database file, so that other processes know not to create it
2) lock the database for writing such that the current process is
   the only process initializing the database.

We use an `EXCLUSIVE` transaction to ensure that "that a write transaction
is started immediately" and "prevent other database connections from
reading the database while the transaction is underway"
(in case we are not already using WAL mode?).
-- https://www.sqlite.org/lang_transaction.html

Since database files seem to only be created when the first instruction is
executed, starting a transaction also doubles as a file creation step (1).