feat: add profile section to the manifest (#1149)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This PR adds a `[profile]` section to the manifest and deprecates the
`hook.script` field. All of the scripts contained in the `profile`
section are sourced by the user's shell, whereas the `hook.on-activate`
script is run in a non-interactive Bash sub-shell.

The `[profile]` table contains 3 fields:
- `profile.common`: this script is sourced by all shells, and therefore
should be written in a shell-agnostic way. It is sourced after
environment variables from the `[vars]` table are set.
- `profile.bash`: this script is sourced after `profile.common` and only
by Bash shells.
- `profile.zsh`: this script is sourced after `profile.common` and only
by Zsh shells.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
- Adds a `[profile]` section to the manifest that contains scripts to be
sourced by the user's shell after activating the environment.
- Deprecates the `hook.script` field.

<!-- Many thanks! -->

---------

Co-authored-by: Bill LeVine <bill@floxdev.com>
Co-authored-by: Yannik Sander <7040031+ysndr@users.noreply.github.com>