fix: allow "flox pull" to overwrite environment with "--force" flag
When using "flox pull --force" while in an environment, it fails to
overwrite the environment. This commit allows overwriting it by
checking if the 'force' flag is set.

To maintain data integrity and avoid potential conflicts between old
and new environments, we delete the old environment before pulling in
the new one.

Closes #1205

Signed-off-by: Charalampos Mitrodimas <charmitro@posteo.net>