feat: stream pkgdb stdout in all cases (#1238)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This PR makes all calls to `pkgdb` stream from `stderr` and `stdout` so
that the pipe buffer between `flox` and `pkgdb` is consistently drained,
and writes from `pkgdb` don't block indefinitely.

This is intended to fix https://github.com/flox/flox/issues/1223.

All calls to `pkgdb` are performed with `pkgdb`'s `stdout` and `stderr`
piped to the calling process (`flox`). For most calls to `pkgdb` we
simply wait for the process to exit, and then read the contents piped
from `stdout` and `stderr` all at once. The major exception to this is
the call to `pkgdb search`, the output of which is streamed in
background threads.

This only works when the output of `pkgdb` is smaller than the buffer
size of the pipe. When a process tries to write to a pipe that has a
full buffer, that write will block. When `pkgdb` makes a large write to
`stdout`, perhaps because there's a very large error coming from Nix, it
cannot make progress until `flox` makes a read from the pipe.

You can stimulate this error by prepending a large string to the output
of the `pkgdb manifest lock` command, which produces the JSON form of
the lockfile:
```cpp
int
LockCommand::run()
{
  nlohmann::json lockfile
    = this->getEnvironment().createLockfile().getLockfileRaw();
  /* Print that bad boii */
  auto bigStr      = std::string( 66000, ' ' );
  auto lockfileStr = lockfile.dump();
  auto output      = bigStr + lockfileStr;
  std::cout << output << '\n';
  return EXIT_SUCCESS;
}
```

Then if you try to install a known-good package with `flox install`
you'll see the program hang, even though the lockfile is valid.
```console
$ flox install ripgrep
<hang>
```

With this fix the install doesn't hang with the large string prepended
to the output.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
Flox should no longer hang in some cases.

<!-- Many thanks! -->