feat(search): Bold search term matches in interactive output
This feature improves the user experience by making it easier to
visually identify the search term matches in the search results when
running `flox search` in an interactive terminal. The bolding is
automatically disabled when the output is redirected to a file or
piped to another command to ensure compatibility with non-interactive
use cases.

Key changes:
    - Added the `console` package for text styles
    - Detect if the output is a terminal (TTY)
    - Updated the `DisplaySearchResults` struct to include a
      `use_bold` field to determine whether to bold the search
      term matches
    - Modified the `Display` trait implementation for
      `DisplaySearchResults` to conditionally bold the package names
      based on the `use_bold` field

Signed-off-by: Charalampos Mitrodimas <charmitro@posteo.net>