test: move expect pull tests to unit tests (#1249)
- Move most `flox pull` integration tests that cover incompatible system
or incompatible package errors to unit tests. 6/8 of the tests moved are
expect tests that are flakey. Unit tests will be faster and more
reliable.
- A few tests that don't use expect were left as integration tests:
  - one that tests a warning printed
- two that test integration with pkgdb, one for a package incompatible
with the current system, and one for an insecure package.
- Two tests for insecure packages were dropped after adding a unit test
checking that building an insecure package returns the expected error.
Testing pull behavior for that case is unecessary since it's identical
to that for packages unavailable for the current system.

Code notes:
- handle_pull_result was factored out from pull_new_environment to allow
testing just the relevant code paths
- push_new_without_building was split out from push_new to allow
creating a ManagedEnvironment that wouldn't build on the current system
- test helpers were added to setup a mock FloxHub that could be pushed
to