feat(search): Bold search term matches in interactive output (#1250)
This feature improves the user experience by making it easier to
visually identify the search term matches in the search results when
running `flox search` in an interactive terminal. The bolding is
automatically disabled when the output is redirected to a file or piped
to another command to ensure compatibility with non-interactive use
cases.

Key changes:
- Added the `atty` dependency to detect if the output is a
      terminal (TTY)
- Updated the `DisplaySearchResults` struct to include a
      `bold_search_term` field to determine whether to bold the search
      term matches
- Added a new `bold_matches` function to apply bolding to the
      search term matches using ANSI escape codes
- Modified the `Display` trait implementation for
      `DisplaySearchResults` to conditionally bold the package names
      based on the `bold_search_term` field

Closes #385

Signed-off-by: Charalampos Mitrodimas <charmitro@posteo.net>