feat: fork watchdog process for flox activate
Investigate idea of forking "watchdog" process for monitoring the
invocation of flox activations for the purpose of performing cleanup
and metrics submission upon completion. This idea may also be useful
for improving the straight-line performance of flox in general,
opening up the possibility of gathering more details from flox
activations, including their duration.