test: port tests for detecting active environment (#1288)
We currently have integration tests checking if we prompt for which
environment to use when there's an activated environment and an
environment in the current directory, and we have a test checking we use
the last activated environment. The tests use expect which is flaky and
slow. Add unit testing coverage instead for each test as follows.

For the tests:
- 'flox install' uses last activated environment
- 'flox install' prompts when an environment is activated and there is
an environment in the current directory
- 'flox install' prompts when an environment is activated and there is
an environment in the containing git repo

Add coverage by:
- Testing `last_activated`
- Factoring out `query_which_environment` from `detect_environment`.
This shows that `detect_environment` and `query_which_environment` are
trivial enough they don't need tests.

For the test:
- show prompts when an environment is activated and there is an
environment in the current directory

Add coverage by:
- Factoring out `manifest_and_lockfile_from_detected_environment` and
unit testing it
- Note: this required adding a test helper for creating a
PathEnvironment, which required factoring out
PathEnvironment::write_new. In the process an unnecessary copy of the
template directory was removed since our template flox env directory
only contains a single file.

---------

Co-authored-by: Yannik Sander <7040031+ysndr@users.noreply.github.com>
Co-authored-by: Zach Mitchell <zmitchell@fastmail.com>