feat: notify when new CLI version is available
Check for a new flox version on downloads.flox.dev, and if one is
available, print a notification to the user.

Only print this notification once every 24 hours.

The notification is printed after command output, and the network call
to check for a new version will timeout after 2 seconds.

If a command fails, the notification will be printed before the error
message.

Errors are not shown to the user, but they are captured with Sentry if
they aren't network or io errors.