feat: notify when new CLI version is available (#1310)
Check for a new flox version on downloads.flox.dev, and if one is
available, print a notification to the user.

Only print this notification once every 24 hours.

The notification is printed after command output, and the network call
to check for a new version will timeout after 2 seconds.
```
flox version 1.0.2

Usage: flox OPTIONS (init|activate|search|install|...) [--help]

Use `flox --help` for full list of commands and more information

First time? Create an environment with `flox init`

🚀  Flox has a new version available. 1.0.2 -> 1.0.3

Get the latest at https://flox.dev/docs/install-flox/
```

If a command fails, the notification will be printed before the error
message:
```
🚀  Flox has a new version available. 1.0.2 -> 1.0.3

Get the latest at https://flox.dev/docs/install-flox/

❌ ERROR: could not parse lockfile: expected value at line 3 column 15
```

The notification is not printed for `flox --version`.

Errors are not shown to the user, but they are captured with Sentry if
they aren't network or IO errors.