fix: multiple FLOX_ENV_LIB_DIRS support for ld-floxlib.so (#1316)
## Proposed Changes

The code to parse multiple FLOX_ENV lib directories was mutating the
FLOX_ENV_LIB_DIRS environment and consequently breaking its ability to
locate libraries in all but the first library on the list.

This patch updates the code to first make a copy of the environment
variable, then split that new string into tokens as it populates a new
array of pointers. To make the code as efficient as possible the string
and array data are maintained in static fixed-length buffers, so we
correspondingly make best efforts not to overflow those buffers in the
process.

## Release Notes

- fixed bug affecting ability to load dynamic shared libraries from
layered environment activations

---------

Co-authored-by: tomberek <tomberek@users.noreply.github.com>