feat: add separate `RawManifest` and `TypedManifest` types
`RawManifest` is a wrapper around a `toml_edit::DocumentMut`
that allows modifications of the raw manifest document,
while preserving comments and user formatting.

`TypedManifest` represents the Manifest data schema for reading/processing of the manifest.
Writing a `TypedManifest` will drop comments and formatting so this should only be used in cases
where these can safely be severed.