refactor: reimplement `InitHook` trait to avoid invalid states caused by `should_run` (#1313)
## Proposed Changes

This PR changes the `InitHook` trait by dropping `should_run` check and
implementing a type-checker friendly design by allowing trait objects to
be instantiated with `None` if they are not intended to run.

The position of the `flox` argument in many methods is also corrected to
be the first argument.
This change could be considered outside of the scope of this PR, and
would drop it / move it to a different PR if requested.

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
Fixes #1266.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->

---------

Co-authored-by: Matthew Kenigsberg <matthew@floxdev.com>