feat: add local environment registry (#1312)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This PR replaces the reverse symlink method of keeping track of managed
environments with a JSON document that keeps track of both path and
managed environments.

The registry is a JSON document with this format:
```
{
  "version": <int>,
  "environments": [
    {
      "hash": <truncated hash(path)>,
      "path": "/path/to/env",
      "envs": [
        {
          "created_at": <timestamp>
          "pointer": <env.json contents>,
        },
        ...
      ]
    },
    ...  
  ]  
}
```
Note that this is slightly different from the format described in
[#1287](https://github.com/flox/flox/issues/1287), the difference being
that `kind` was dropped and `pointer` was made to always contain a
pointer. This makes more sense because it no longer allows
inconsistencies where `kind` could be `"managed"` but `pointer` could be
`null`. It also allows us to discern whether two path environments are
the same because the pointer stores the name.

The semantics for the registry are as follows:
- Environments are registered when
    - A path environment is created via `init`
    - A path environment is opened
- This covers the case where a user creates an environment through
`init` or `pull`, manually deletes the `.flox` directory, then does a
`git pull`, which places a `.flox` directory in the project without
going through any `flox` machinery.
    - A managed environment is opened (which happens during a pull, etc)
- When a managed environment is opened, it is registered.
- "Same environment" is defined as
    - At the same location (identified by the hash of its path)
- With the same environment pointer (e.g. both are `PathPointer`s with
the same `name`)
- When registering an environment at a particular location
- If the latest registered environment is the same as the environment
being registered, this is a no-op.
- If the latest registered environment is _not_ the same, the
environment is added to the list of environments registered to this
location.
- When deleting an environment
- If the latest environment at a particular location is the same as the
environment being deleted, the environment is _unregistered_ and removed
from the registry.
- If the latest environment at the location is _not_ the same as the one
being deleted, an error is raised. (**Note: I'm not sure if this is the
best behavior, I'm open to suggestions.**)
- If there are no more registered environments at a location, the entry
for this location is removed entirely.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->

---------

Co-authored-by: Matthew Kenigsberg <matthew@floxdev.com>
Co-authored-by: Yannik Sander <7040031+ysndr@users.noreply.github.com>