feat: add resolve shim and catalog feature flag (#1328)
- Add a wrapper client around the autogenerated client. This client can
be used throughout flox-rust-sdk. It uses redefined types that are very
similar to the autogenerated types, but there are some slight
differences (currently just system and stability). These redefined types
can be used throughout flox-rust-sdk, but they'll give us a single place
to manage changes.
- Add a feature flag to use the catalog, and add an optional
catalog_client to the Flox struct dependant on the feature flag
- Add an environment variable to control whether the actual catalog
client is used or a mock client, which is left unimplemented!()

---------

Co-authored-by: Yannik Sander <yannik@floxdev.com>
Co-authored-by: Yannik Sander <7040031+ysndr@users.noreply.github.com>