fix: don't notify of updates for dev builds
Currently if FLOX_SENTRY_ENV isn't set, we default to checking the
stable channel for a version. This is then compared for equality with
FLOX_VERSION. For dev builds, this will return false, so an update
notification is printed.

Dev builds don't have semantic versions (e.g. 1.0.3-g294e510) so we
can't use something other than equality to compare with latest.

Instead, just don't print the notification for dev builds.