fix: update manifest template and docs per #1299
This PR has kicked off a discussion regarding the best manifest template and
documentation that is somewhat outside of the scope of the effort. Of course
it is absolutely essential that we make this template as clear as possible,
but the job of perfecting the wording will never be complete and it's not
strictly necessary to achieve perfection as part of this effort. With that
I'm applying this patch which ought to move us closer to something we all
agree upon, but we'll soon call time and continue this in a separate PR so
that this one can be merged ASAP.

Details of the full conversation are available here:
https://github.com/flox/flox/pull/1299/files#r1570842930