fix: adding comment detailing the activate:hook tests
The activate:hook test has become a full enumeration of all possible
activation scenarios along with tests to verify that the expected
activation scripts run in each instance, and for each supported shell.
Adding a comment to explain the [4] scenarios in more detail, which
should help the reader to understand why it is that certain strings
are and are not expected to appear.

This patch also fixes the bug that the hook on-activate script wasn't
being run and cleans up the ordering of assert/refute statements to
make more sense.