refactor: wip Test handle_pull_result messages
I've tried using an enum because we need to know the action and content
of the message. We don't have the emoji formatting though because that's
applied later.

Things I'm unsure about:

- using String when the existing methods use a trait?
- not able to do partial matches on the contents
- there are some intermediate warnings that aren't captured

An alternative could be to construct something like a logger that will
accept a different output buffer for the tests. It's a more invasive
change though.