feat: nix-based floxenv builder
This commit adds the `flox-buildenv` package, which is an adaptation of the
upstream nixpkgs `buildenv` builder altered to produce multiple outputs:

* `runtime`: the same env as created by nixpkgs `buildenv` upstream
* `develop`: as above, but recurses to include links for all requisite packages
* `build-<pname>`: for each manifest build, creates env referring only to
  the "toplevel" packages, optionally filtered by `runtime-packages`

The Flox environment builder also includes logic to automatically add the
following packages to each output:

* the flox "interpreter": also known as "etc-profiles", this package includes
  scripts and requisites required to activate a flox environment
* the "manifest" package: this package contains `manifest.lock` and other files
  derived from that data (e.g. the `activate.d/envrc` file)

The bulk of the business logic resides in the `buildenv.nix` and `builder.pl`
files, but the buildenv package also includes a `buildenv.bash` script used for
testing and available for use as we transition from invoking `pkgdb buildenv`
to `${flox-buildenv}/bin/buildenv`. Ultimately, the plan is to deprecate this
script and call `nix` directly from Rust for realising packages and building
the resulting environments.