refactor: spinners and messaging during `flox pull` (#1337)
* Add missing spinners when pulling new environments that require the
current system to be added to the manifest
(https://github.com/flox/flox/issue/1301) Rebuilding of the modified
environment was not covered by any spinner.
* Change messages when pulling results in a broken environment. Previous
messages always implied that the environment was ready to be activated
even when we knew it was broken.
* Refactor pulling existing environments to match the structure of
pulling new environments. Mainly moving messaging inwards.

Co-Authored-By: dan.carley@gmail.com

---------

Co-authored-by: Dan Carley <dan.carley@gmail.com>