fix: remove NIX_PKG_BIN, backport fixes from fish port
Discovered a few bugs while porting fish, notably the failure to properly
quote and add semicolons between statements in eval'd strings. This patch
back-ports those fixes to the issue-1187 branch to be incorporated into
PR#1299. It also removes mention of NIX_PKG_BIN now that we are no longer
using it in the pkgdb path.