feat: add `flox envs` command (#1264)
This PR adds an environment registry and `flox envs` command.

Flox commands has so far been mostly stateless, i.e. the fact that an
environment exists at a given path is "forgotten" as soon as the command
completes.
`flox envs` is supposed to list both active and _available/inactive_
environments (that are represented by .flox directories).
To avoid _scanning_ for `.flox` directories, we use the env registry
introduced in #1312.

In addition, the `flox envs` supports filtering the output to only
active environments as well as formatting its output as JSON.