feat: mock client returns mock data (#1346)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
Creates a `MockClient` that can be seeded with responses in unit tests
or read responses from a JSON document on disk.

The JSON document on disk is used when the `_FLOX_USE_CATALOG_MOCK`
environment variable is set to the path of the JSON document. The JSON
document must be an array where each element is a single response.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->