feat: add proptest for manifest (#1365)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
Derives the `proptest::Arbitrary` trait for the new manifest types. It's
likely that the generated data will need to be tuned to ensure it makes
sense (invariants upheld, etc) and that test case generation isn't too
slow. For performance concerns, see the tweaks made to how data is
generated for the registry.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A


<!-- Many thanks! -->