build: use fenix provided nightly rustfmt
We depend on several nightly features of rustfmt,
so pick the current nightly version.
We're using `default.withComponents`
which _should_ only pull the nightly rustfmt component.
Alternatively, we could use nixpkgs.rustfmt,
and rebuild with a (stable) fenix toolchain and `asNightly = true`,
which would avoid the need to pull another channel altogether.