fix: ensure zsh profile scripts run only once
Eagle-eyed Matthew spotted that profile scripts were being invoked
twice for zsh on account of the fact that we were sourcing the flox
init script from each of .zshenv and .zshrc. Digging into it further
realised that we should not be sourcing from .zshenv in the interactive
case because the .zshrc will source it again.

Added a test to verify that we're only sourcing hook and profile scripts
once for each of bash and zsh, and in each of the interactive and
non-interactive cases.