fix: split InitCustomization() scripts into individual strings
With the activate refactor we need to be able to configure the various
profile.* scripts independently of hook.on_activate. This patch factors
these strings into separate struct elements and updates the logic to
recombine them into template manifests accordingly.

Also renaming the "script" variable to be the more explicit
"hook_on_activate" struct member to make this all more clear.

BROKEN: the init test is known not to work, passing off to Matthew to fix.